rename mutable to mutation to be parallel with reduction
